[core] Move OnlineFileSource to platform
Preparation for OfflineFileSource which depends on SQLite
