+++ gyp/platform-ios.gypi
@@ -13,6 +13,8 @@
         '../platform/default/async_task.cpp',
         '../platform/default/run_loop.cpp',
         '../platform/default/timer.cpp',
+        '../platform/default/default_file_source.cpp',
+        '../platform/default/online_file_source.cpp',
         '../platform/darwin/log_nslog.mm',
         '../platform/darwin/string_nsstring.mm',
         '../platform/darwin/application_root.mm',
